backend/requirements.txt
Django==5.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
redis==5.0.1
pika==1.3.2
elasticsearch==8.11.0
Pillow==10.1.0
gunicorn==21.2.0
